Hasta aqui todo correcto, clase Cliente se cambia el tipo de celular

+++ src/main/java/com/platzi/market/persistence/entity/Cliente.java
@@ -15,13 +15,9 @@ public class Cliente {
 
     @Id
     private String id;
-
     private String nombre;
-
     private String apellidos;
-
-    private Integer celular;
-
+    private Long celular;
     private String direccion;
 
     @Column(name = "correo_electronico")
